Add onSuccess and onFailure methods to Result classes
The onSuccess and onFailure methods were added to both the Result and the Result<T> classes. These methods are engineered to invoke specified actions depending on the success or failure of the operation. Corresponding unit tests were also created to check the proper functioning of these new methods.

Took 32 minutes