"Implemented guest event participation and invitation acceptance
The code now allows for a guest to participate in an event and to accept a invitation. Participation validation checks have been added to ensure that an event is active, not full and not past. For invitation acceptance, duplications are prevented and the status is updated in the system accordingly. Tests for the new implementations are also added."

Took 2 hours 43 minutes